refactor(release): split release.rs into release/ module
Purely mechanical refactor — no logic changes. Extracts the 1,740-line
src/release.rs into src/release/mod.rs plus submodules: version.rs,
bump.rs, toml_utils.rs, changelog.rs, git.rs, and tests.rs. Public API
(ReleaseOptions, run()) remains importable from crate::release:: unchanged.
